DEV: updated histogram functions for image processing in VisTrails
